backend/requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
httpx
minio
ipfshttpclient
geoalchemy2
cryptography
pytest